chore(quality): execute CLEANUP P0+P1 — dead edges, god-file splits, hidden-coupling extractions
  10 dead Cargo edges cut; dead cfg blocks removed; lock audit (23 sites, all
  tests); sqlite.rs and retrieval.rs split under the 1,200 ceiling;
  plan_grammar/synthesis_common/middleware-shared extractions make git-observed
  hidden coupling structural; sovereign-contracts docs 798→0; clippy --fix
  sweep (baseline 2,555→1,637); api snapshots for the contract leaves
  (hub crates parked on a nightly ICE in lance-index). Fixes SYSTEM_OVERVIEW's
  fictional planner-annotation grammar. All gates green.

+++ sovereign/crates/sovereign-cli-llm/src/meshapp_cmd.rs
@@ -322,7 +322,7 @@ async fn op_handler(
             ))
         }),
         "document_feed" => {
-            let limit = a.limit.unwrap_or(14).clamp(1, 90) as usize;
+            let limit = a.limit.unwrap_or(14).clamp(1, 90);
             sovereign_meshapp::document_feed(idx, limit)
                 .await
                 .map(to_val)
